requirements.txt
anyio==4.2.0
attrs==23.1.0
boto3==1.15.2
botocore==1.18.18
bson==0.5.10
cassandra-driver==3.28.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
Deprecated==1.2.14
dnspython==2.4.2
exceptiongroup==1.2.0
fastapi==0.99.0
fastjsonschema==2.18.1
geomet==0.2.1.post1
h11==0.14.0
idna==3.6
jmespath==0.10.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jwcrypto==1.5.0
mangum==0.17.0
motor==3.3.2
passlib==1.7.4
pillow==10.2.0
pycparser==2.21
pydantic==1.10.13
PyJWT==2.8.0
pymongo==4.6.1
python-dateutil==2.8.2
python-jwt==4.1.0
referencing==0.32.0
rpds-py==0.15.2
s3transfer==0.3.7
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
svgwrite==1.4.3
Tree==0.2.4
typing_extensions==4.9.0
urllib3==1.25.11
uvicorn==0.24.0.post1
wrapt==1.16.0
